Базовая установка и настройка линтеров
