+++ .eslintrc.js
@@ -1,0 +1,22 @@
+module.exports = {
+    env: {
+        browser: true,
+        es2021: true
+    },
+    extends: ["plugin:react/recommended", "airbnb", "plugin:prettier/recommended"],
+    overrides: [],
+    parserOptions: {
+        ecmaVersion: "latest",
+        sourceType: "module"
+    },
+    plugins: ["react"],
+    rules: {
+        "react/react-in-jsx-scope": "off",
+        "prettier/prettier": [
+            "error",
+            {
+                endOfLine: "auto"
+            }
+        ]
+    }
+};
