+++ .prettierrc
@@ -1,0 +1,9 @@
+{
+  "semi": true,
+  "tabWidth": 4,
+  "printWidth": 120,
+  "singleQuote": false,
+  "trailingComma": "none",
+  "jsxBracketSameLine": true,
+  "endOfLine": "auto"
+}
