--- src/App.js
@@ -1,7 +1,0 @@
-function App() {
-  return (
-    <h1>Hi, Gleb!</h1>
-  );
-}
-
-export default App;
