+++ src/App.jsx
@@ -1,0 +1,10 @@
+/**
+ * Главный компонент приложения
+ * @returns {JSX.Element}
+ * @constructor
+ */
+function App() {
+    return <h1>Hi, Gleb!</h1>;
+}
+
+export default App;
